requirements.txt
pytest
allure-pytest
python-dotenv
PyYAML
requests
pytest-html
pytest-sugar
